requirements.txt
streamlit
pandas
SQLAlchemy
psycopg2-binary
openpyxl
openai
python-dotenv
